+++ backend-rust/src/migrations.rs
@@ -1,4 +1,5 @@
 use anyhow::Context;
+use concordium_rust_sdk::v2;
 use sqlx::{Executor, PgPool};
 use std::cmp::Ordering;
 use tokio_stream::StreamExt;
@@ -7,6 +8,9 @@ use tracing::info;
 use std::{str::FromStr};
 
 type Transaction = sqlx::Transaction<'static, sqlx::Postgres>;
+
+mod m0005_fix_dangling_delegators;
+mod m0006_canonical_address_and_transaction_search_index;
 
 /// Ensure the current database schema version is compatible with the supported
 /// schema version.
@@ -76,7 +80,11 @@ The following breaking schema migrations have happened:
 }
 
 /// Migrate the database schema to the latest version.
-pub async fn run_migrations(pool: PgPool, cancel_token: CancellationToken) -> anyhow::Result<()> {
+pub async fn run_migrations(
+    pool: PgPool,
+    endpoints: Vec<v2::Endpoint>,
+    cancel_token: CancellationToken,
+) -> anyhow::Result<()> {
     cancel_token
         .run_until_cancelled(async move {
             ensure_migrations_table(&pool).await?;
@@ -85,7 +93,7 @@ pub async fn run_migrations(pool: PgPool, cancel_token: CancellationToken) -> an
             info!("Latest database schema version {}", SchemaVersion::LATEST.as_i64());
             while current < SchemaVersion::LATEST {
                 info!("Running migration from database schema version {}", current.as_i64());
-                let new_version = current.migration_to_next(&pool).await?;
+                let new_version = current.migration_to_next(&pool, endpoints.as_slice()).await?;
                 info!("Migrated database schema to version {} successfully", new_version.as_i64());
                 current = new_version
             }
@@ -177,6 +185,8 @@ pub enum SchemaVersion {
     PayDayPoolCommissionRates,
     #[display("0004:PayDayLotteryPowers")]
     PayDayLotteryPowers,
+    #[display("0005:Fix invalid data of dangling delegators.")]
+    FixDanglingDelegators,
     #[display("0006:AccountBaseAddress")]
     AccountBaseAddress,
 }
@@ -184,7 +194,7 @@ impl SchemaVersion {
     /// The minimum supported database schema version for the API.
     /// Fails at startup if any breaking database schema versions have been
     /// introduced since this version.
-    pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion = SchemaVersion::AccountBaseAddress;
+    pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion = SchemaVersion::FixDanglingDelegators;
     /// The latest known version of the schema.
     const LATEST: SchemaVersion = SchemaVersion::AccountBaseAddress;
 
@@ -206,12 +216,17 @@ impl SchemaVersion {
             SchemaVersion::IndexBlocksWithNoCumulativeFinTime => false,
             SchemaVersion::PayDayPoolCommissionRates => false,
             SchemaVersion::PayDayLotteryPowers => false,
+            SchemaVersion::FixDanglingDelegators => false,
             SchemaVersion::AccountBaseAddress => false,
         }
     }
 
     /// Run migrations for this schema version to the next.
-    async fn migration_to_next(&self, pool: &PgPool) -> anyhow::Result<SchemaVersion> {
+    async fn migration_to_next(
+        &self,
+        pool: &PgPool,
+        endpoints: &[v2::Endpoint],
+    ) -> anyhow::Result<SchemaVersion> {
         let mut tx = pool.begin().await?;
         let start_time = chrono::Utc::now();
         let new_version = match self {
@@ -238,30 +253,18 @@ impl SchemaVersion {
                 tx.as_mut().execute(sqlx::raw_sql(include_str!("./migrations/m0004.sql"))).await?;
                 SchemaVersion::PayDayLotteryPowers
             }
-            SchemaVersion::PayDayLotteryPowers => unimplemented!(
+            SchemaVersion::PayDayLotteryPowers => {
+                m0005_fix_dangling_delegators::run(&mut tx, endpoints).await?;
+                SchemaVersion::FixDanglingDelegators
+            }
+            SchemaVersion::FixDanglingDelegators => {
+                m0006_canonical_address_and_transaction_search_index::run(&mut tx).await?;
+                SchemaVersion::AccountBaseAddress
+            },
+            SchemaVersion::AccountBaseAddress => unimplemented!(
                 "No migration implemented for database schema version {}",
                 self.as_i64()
-            ),
-            SchemaVersion::AccountBaseAddress => {
-//                let tx = tx.as_mut();
-//                tx.execute(sqlx::raw_sql(include_str!("./migrations/m0006.sql"))).await?;
-//                let mut accounts = sqlx::query!("SELECT index, address FROM accounts").fetch(pool);
-//                while let Some(account) = accounts.try_next().await? {
-//                    let account_address = concordium_rust_sdk::base::contracts_common::AccountAddress::from_str(&account.address)?;
-//                    let canonical_account_address = account_address.get_canonical_address();
-//                    sqlx::query!(
-//                        "UPDATE accounts
-//                            SET canonical_address = $2
-//                            WHERE index = $1",
-//                        account.index,
-//                        canonical_account_address.0.into()
-//                    )
-//                    .execute(tx.as_mut())
-//                    .await?;
-//                }
-//
-                SchemaVersion::AccountBaseAddress
-            }
+            )
         };
         let end_time = chrono::Utc::now();
         insert_migration(&mut tx, &new_version.into(), start_time, end_time).await?;